fix: remove gradient from typography variants and rename createpage
