+++ packages/hash/frontend/src/components/Modals/CreatePageModal/CreatePage.module.scss
